Refactor leveldb database access
